refactor: extract rule-module tests; drop pytest runtime coupling (PR-08)
Each per-dataset rule module (src/pdsfile/pds{3,4}file/rules/*.py) carried its
unit tests inline behind `import pytest` + `from .pytest_support import *`, which
made pytest a runtime dependency of `import pdsfile`. PR-08 extracts those tests
verbatim to the top-level tests/ tree and removes the coupling. No test logic
and no goldens changed; the additive-coverage step remains deferred.

- Extract 13 pds3 + 3 pds4 rule-module test sections to
  tests/rules/pds{3,4}/test_<dataset>.py with explicit import headers (no star
  imports; module-internal translator tables imported from the rule module).
- Merge the two rules/pytest_support.py into tests/rules/support.py; move
  pdsfile_test_helper.py to tests/support/ (relative import made absolute).
- Strip the test sections from the rule modules; add `import os` to COVIMS_0xxx
  (production os.path use previously satisfied by the star import); remove the
  dead pytest scaffolding from cassini_iss / cassini_vims (no live tests).
- Move conftest.py to tests/conftest.py; api-freeze now uses
  --confcutdir=tests/api to stay hermetic.
- pyproject: drop pytest from [project].dependencies (stays in the dev extra);
  add repo root to pytest pythonpath so the tests.* absolute imports resolve;
  regenerate the ruff ratchet (rule-module entries shrink; new test files added).
- CI: repoint the rule-module globs to tests/rules/pds{3,4}/.
- New permanent clean-install gate (scripts/clean_install_check.sh +
  check_runtime_imports.py, wired into run-all-checks.sh and CI): installs the
  project with no extras and imports the full module surface, so a runtime
  dependency leak (e.g. a stray pytest import) fails the build.
- api-freeze: add the pre-approved category #2 forgiveness (rule modules lose
  their test-only public names) to manifest_allowlist.json.
- Document the layout and deferred additive-coverage applicability in
  tests/rules/README.md.

Verified against the limited holdings: --mode ns 679 passed / 34 skipped,
--mode s 555 passed / 3 skipped (unchanged counts); run-all-checks code gates
(ruff, pyroma 10/10, api-freeze, clean-install) green.
