+++ scripts/automated_tests/pdsfile_main_test.sh
@@ -19,12 +19,29 @@ pip3 install -e ".[dev]"
 echo
 
 echo "================================================================"
+echo "CLEAN-INSTALL GATE (runtime-dependency leak)"
+echo "================================================================"
+echo
+# PR-08: verify the full pdsfile module surface imports from a bare `pip install
+# .` (no extras). Catches a runtime module still importing a dev-only dependency
+# (e.g. pytest). Holdings-independent; runs before the data tests so a leak fails
+# fast. Builds its own throwaway venv.
+bash scripts/clean_install_check.sh
+if [ $? -ne 0 ]; then
+    echo "**************************************"
+    echo "*** CLEAN-INSTALL GATE FAILED      ***"
+    echo "**************************************"
+    exit -1
+fi
+echo
+
+echo "================================================================"
 echo "PDSFILE NOT-SHELVES-ONLY TESTS"
 echo "================================================================"
 echo
 echo "Test start:" `date`
 echo
-python -m coverage run -m pytest tests/api/ tests/pds3file/ src/pdsfile/pds3file/rules/*.py tests/pds4file/ src/pdsfile/pds4file/rules/*.py --mode ns
+python -m coverage run -m pytest tests/api/ tests/pds3file/ tests/rules/pds3/ tests/pds4file/ tests/rules/pds4/ --mode ns
 if [ $? -ne 0 ]; then
     echo "**************************************************"
     echo "*** PDSFILE NOT-SHELVES-ONLY FAILED UNIT TESTS ***"
@@ -43,7 +60,7 @@ echo "================================================================"
 echo
 echo "Test start:" `date`
 echo
-python -m coverage run -a -m pytest tests/pds3file/ src/pdsfile/pds3file/rules/*.py --mode s
+python -m coverage run -a -m pytest tests/pds3file/ tests/rules/pds3/ --mode s
 if [ $? -ne 0 ]; then
     echo "********************************************"
     echo "*** PDSFILE SHELVES-ONLY FAILED UNIT TESTS ***"
