+++ scripts/run-all-checks.sh
@@ -22,6 +22,7 @@
 #   --pytest               Run pytest only
 #   --pyroma               Run pyroma only
 #   --api-freeze           Run the public-API freeze check only
+#   --clean-install        Run the clean-install (runtime-dep leak) gate only
 #   --bandit               Run bandit only
 #   --vulture              Run vulture only
 #   --sphinx               Run Sphinx build only
@@ -37,7 +38,7 @@
 #
 #   RUN_* (set by this script from CLI or full-run defaults): RUN_RUFF_CHECK,
 #   RUN_RUFF_FORMAT, RUN_MYPY, RUN_PYTEST, RUN_PYROMA, RUN_API_FREEZE,
-#   RUN_BANDIT, RUN_VULTURE, RUN_SPHINX, RUN_PYMARKDOWN
+#   RUN_CLEAN_INSTALL, RUN_BANDIT, RUN_VULTURE, RUN_SPHINX, RUN_PYMARKDOWN
 #
 #   Per-check toggles (true/false). rms-pdsfile turns gates on as they become
 #   able to pass (see pdsfile_overrides.mdc and the modernization plan). Each
@@ -49,6 +50,7 @@
 #     ENABLE_PYTEST       (default: false — hermetic pytest arrives Phase 4)
 #     ENABLE_PYROMA       (default: true)
 #     ENABLE_API_FREEZE   public-API freeze (default: true)
+#     ENABLE_CLEAN_INSTALL clean-install runtime-dep leak gate (default: true)
 #     ENABLE_BANDIT       (default: false — never)
 #     ENABLE_VULTURE      (default: false — never)
 #     ENABLE_SPHINX       (default: false — docs phase)
@@ -84,6 +86,7 @@ RUN_MYPY=false
 RUN_PYTEST=false
 RUN_PYROMA=false
 RUN_API_FREEZE=false
+RUN_CLEAN_INSTALL=false
 RUN_BANDIT=false
 RUN_VULTURE=false
 RUN_SPHINX=false
@@ -104,6 +107,7 @@ SCOPE_SPECIFIED=false
 : "${ENABLE_PYTEST:=false}"
 : "${ENABLE_PYROMA:=true}"
 : "${ENABLE_API_FREEZE:=true}"
+: "${ENABLE_CLEAN_INSTALL:=true}"
 : "${ENABLE_BANDIT:=false}"
 : "${ENABLE_VULTURE:=false}"
 : "${ENABLE_SPHINX:=false}"
@@ -223,6 +227,7 @@ while [[ $# -gt 0 ]]; do
             RUN_PYTEST=true
             RUN_PYROMA=true
             RUN_API_FREEZE=true
+            RUN_CLEAN_INSTALL=true
             RUN_BANDIT=true
             RUN_VULTURE=true
             SCOPE_SPECIFIED=true
@@ -269,6 +274,11 @@ while [[ $# -gt 0 ]]; do
             SCOPE_SPECIFIED=true
             shift
             ;;
+        --clean-install)
+            RUN_CLEAN_INSTALL=true
+            SCOPE_SPECIFIED=true
+            shift
+            ;;
         --bandit)
             RUN_BANDIT=true
             SCOPE_SPECIFIED=true
@@ -309,6 +319,7 @@ if [ "$SCOPE_SPECIFIED" = false ]; then
     RUN_PYTEST=true
     RUN_PYROMA=true
     RUN_API_FREEZE=true
+    RUN_CLEAN_INSTALL=true
     RUN_BANDIT=true
     RUN_VULTURE=true
     RUN_SPHINX=true
@@ -336,6 +347,7 @@ _code_checks_any_scheduled() {
     [ "$RUN_PYTEST" = true ] && [ "$ENABLE_PYTEST" = true ] && return 0
     [ "$RUN_PYROMA" = true ] && [ "$ENABLE_PYROMA" = true ] && return 0
     [ "$RUN_API_FREEZE" = true ] && [ "$ENABLE_API_FREEZE" = true ] && return 0
+    [ "$RUN_CLEAN_INSTALL" = true ] && [ "$ENABLE_CLEAN_INSTALL" = true ] && return 0
     [ "$RUN_BANDIT" = true ] && [ "$ENABLE_BANDIT" = true ] && return 0
     [ "$RUN_VULTURE" = true ] && [ "$ENABLE_VULTURE" = true ] && return 0
     return 1
@@ -372,9 +384,9 @@ run_code_checks() {
     local failed_checks=""
 
     # rms-pdsfile: lint the package under src/pdsfile, the top-level tests/ tree
-    # (tests moved out of the package in PR-07), the standalone scripts, and the
-    # root conftest.
-    RUFF_TARGETS="src/pdsfile tests scripts conftest.py"
+    # (tests moved out of the package in PR-07; conftest.py moved into tests/ in
+    # PR-08), and the standalone scripts.
+    RUFF_TARGETS="src/pdsfile tests scripts"
 
     if [ "$RUN_RUFF_CHECK" = true ] && [ "$ENABLE_RUFF_CHECK" = true ]; then
         print_info "Running ruff check..."
@@ -436,18 +448,35 @@ run_code_checks() {
         fi
     fi
 
-    # rms-pdsfile: public-API freeze. --confcutdir=tests bypasses the root
-    # conftest.py (which still requires holdings env vars to import until PR-09),
-    # so this check is hermetic. The freeze test regenerates the manifest in a
-    # clean subprocess and needs no holdings itself.
+    # rms-pdsfile: public-API freeze. --confcutdir=tests/api bypasses the
+    # tests/conftest.py (PR-08 moved conftest.py from the repo root into tests/;
+    # it still requires holdings env vars to import until PR-09), so this check
+    # stays hermetic. The freeze test regenerates the manifest in a clean
+    # subprocess and needs no holdings itself.
     if [ "$RUN_API_FREEZE" = true ] && [ "$ENABLE_API_FREEZE" = true ]; then
         print_info "Running API-freeze check..."
-        if python -m pytest tests/api/test_api_freeze.py --confcutdir=tests -p no:cacheprovider -q; then
+        if python -m pytest tests/api/test_api_freeze.py --confcutdir=tests/api -p no:cacheprovider -q; then
             print_success "API-freeze check passed"
         else
             print_error "API-freeze check failed"
             failed=true
             failed_checks="${failed_checks}Code - API freeze"$'\n'
+        fi
+    fi
+
+    # rms-pdsfile: clean-install gate (PR-08, permanent). Builds a throwaway
+    # venv, installs the project with NO extras, and imports the whole public
+    # module surface. The only check that catches a runtime-dependency leak
+    # (every other run has the dev extra present). Runs its own subprocess venv,
+    # independent of the activated dev venv here.
+    if [ "$RUN_CLEAN_INSTALL" = true ] && [ "$ENABLE_CLEAN_INSTALL" = true ]; then
+        print_info "Running clean-install gate (runtime-dep leak)..."
+        if bash "$PROJECT_ROOT/scripts/clean_install_check.sh"; then
+            print_success "Clean-install gate passed"
+        else
+            print_error "Clean-install gate failed"
+            failed=true
+            failed_checks="${failed_checks}Code - Clean install"$'\n'
         fi
     fi
