+++ src/pdsfile/pds3file/rules/EBROCC_xxxx.py
@@ -231,90 +231,3 @@ pds3file.Pds3File.FILESPEC_TO_BUNDLESET = filespec_to_bundleset + pds3file.Pds3F
 ##########################################################################################
 
 pds3file.Pds3File.SUBCLASSES['EBROCC_xxxx'] = EBROCC_xxxx
-
-##########################################################################################
-# Unit tests
-##########################################################################################
-
-import pytest
-from .pytest_support import *
-
-@pytest.mark.parametrize(
-# Allow duplicated '/volumes/EBROCC_xxxx/EBROCC_0001/BROWSE/ESO1M/ES1_EGB.LBL'
-# and '/volumes/EBROCC_xxxx/EBROCC_0001/BROWSE/ESO1M/ES1_EPB.LBL' here. OPUS
-# will ignore the duplicated items
-    'input_path,expected',
-    [
-        ('volumes/EBROCC_xxxx/EBROCC_0001/DATA/ESO1M/ES1_EPD.TAB',
-         'EBROCC_xxxx/opus_products/ES1_EPD.txt')
-    ]
-)
-def test_opus_products(request, input_path, expected):
-    update = request.config.option.update
-    opus_products_test(pds3file.Pds3File, input_path, TEST_RESULTS_DIR+expected, update)
-
-@pytest.mark.parametrize(
-    'input_path,category,expected',
-    [
-        ('volumes/EBROCC_xxxx/EBROCC_0001/DATA/ESO1M/ES1_EPD.TAB',
-         'volumes',
-         'EBROCC_xxxx/associated_abspaths/ES1_EPD.txt')
-    ]
-)
-def test_associated_abspaths(request, input_path, category, expected):
-    update = request.config.option.update
-    associated_abspaths_test(pds3file.Pds3File, input_path, category,
-                             TEST_RESULTS_DIR+expected, update)
-
-def test_opus_id_to_primary_logical_path():
-    TESTS = [
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/ESO1M/ES1_IPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/ESO1M/ES1_EPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/ESO22M/ES2_IPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/MCD27M/MCD_IPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/IRTF/IRT_IPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/LICK1M/LIC_EPD.TAB',
-        'volumes/EBROCC_xxxx/EBROCC_0001/DATA/PAL200/PAL_IPD.TAB',
-    ]
-
-    for logical_path in TESTS:
-        test_pdsf = pds3file.Pds3File.from_logical_path(logical_path)
-        opus_id = test_pdsf.opus_id
-        opus_id_pdsf = pds3file.Pds3File.from_opus_id(opus_id)
-        assert opus_id_pdsf.logical_path == logical_path
-
-        # Gather all the associated OPUS products
-        product_dict = test_pdsf.opus_products()
-        product_pdsfiles = []
-        for pdsf_lists in product_dict.values():
-            for pdsf_list in pdsf_lists:
-                product_pdsfiles += pdsf_list
-
-        # Filter out the metadata/documents products and format files
-        product_pdsfiles = [pdsf for pdsf in product_pdsfiles
-                                 if pdsf.voltype_ != 'metadata/'
-                                 and pdsf.voltype_ != 'documents/']
-        product_pdsfiles = [pdsf for pdsf in product_pdsfiles
-                                 if pdsf.extension.lower() != '.fmt']
-
-        # Gather the set of absolute paths
-        opus_id_abspaths = set()
-        for pdsf in product_pdsfiles:
-            opus_id_abspaths.add(pdsf.abspath)
-
-        for pdsf in product_pdsfiles:
-            # Every version is in the product set
-            for version_pdsf in pdsf.all_versions().values():
-                assert version_pdsf.abspath in opus_id_abspaths
-
-            # Every viewset is in the product set
-            for viewset in pdsf.all_viewsets.values():
-                for viewable in viewset.viewables:
-                    assert viewable.abspath in opus_id_abspaths
-
-            # Every associated product is in the product set except metadata
-            for category in ('volumes', 'previews', 'diagrams'):
-                for abspath in pdsf.associated_abspaths(category):
-                    assert abspath in opus_id_abspaths
-
-##########################################################################################
